+++ src/tirx/op/op.cc
@@ -44,12 +44,12 @@ using namespace tirx;
 // macro to register an unary op
 #define TVM_TIR_REGISTER_PURE_UNARY_OP(OpName)                             \
   TVM_TIR_REGISTER_OP(OpName).set_num_inputs(1).set_attr<TCallEffectKind>( \
-      "TCallEffectKind", Integer(CallEffectKind::kPure))
+      "TCallEffectKind", static_cast<int64_t>(CallEffectKind::kPure))
 
 // macro to register an binary op
 #define TVM_TIR_REGISTER_PURE_BINARY_OP(OpName)                            \
   TVM_TIR_REGISTER_OP(OpName).set_num_inputs(2).set_attr<TCallEffectKind>( \
-      "TCallEffectKind", Integer(CallEffectKind::kPure))
+      "TCallEffectKind", static_cast<int64_t>(CallEffectKind::kPure))
 
 runtime::DataType GetRuntimeDataType(const Type& type) {
   if (auto* n = type.as<PrimTypeNode>()) {
@@ -1167,12 +1167,12 @@ TVM_TIR_REGISTER_PURE_BINARY_OP("ldexp");
 TVM_TIR_REGISTER_OP("TVMBackendAllocWorkspace")
     .set_num_inputs(5)
     .set_attr<TGlobalSymbol>("TGlobalSymbol", "TVMBackendAllocWorkspace")
-    .set_attr<TCallEffectKind>("TCallEffectKind", Integer(CallEffectKind::kOpaque));
+    .set_attr<TCallEffectKind>("TCallEffectKind", static_cast<int64_t>(CallEffectKind::kOpaque));
 
 TVM_TIR_REGISTER_OP("TVMBackendFreeWorkspace")
     .set_num_inputs(3)
     .set_attr<TGlobalSymbol>("TGlobalSymbol", "TVMBackendFreeWorkspace")
-    .set_attr<TCallEffectKind>("TCallEffectKind", Integer(CallEffectKind::kOpaque));
+    .set_attr<TCallEffectKind>("TCallEffectKind", static_cast<int64_t>(CallEffectKind::kOpaque));
 
 // expose basic functions to node namespace
 TVM_FFI_STATIC_INIT_BLOCK() {
